requirements.txt
huggingface_hub==0.17.3
librosa==0.10.2.post1
matplotlib==3.8.0
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.1
scikit_learn==1.3.0
soundfile==0.12.1
thread==2.0.5
torch==2.0.1+cu117
torchvision==0.15.2+cu117
torchaudio==2.0.2
tqdm==4.65.0
git+https://github.com/wenet-e2e/wespeaker.git
